requirements.txt
pyenchant
regex
deep_translator
elevenlabs